Using elementId in dropdown instead of element.id
It fixes null _dropdownContentId which causes all dropdowns on the page
having the same menu items.

I'm not sure what the root cause of the problem, but it looks like
component's `element` is null when it's initializing. Using `elementId`
fixes it.

I'm seing this problem on a blank new Ember 1.13.1 project.