Enable selection property on {{md-select}}
And pass it through to the Ember select view. This allows the select to update belongsTo/hasMany attributes on the model.

Couldn't figure out how to run the in repo tests, but I've anecdotally confirmed it works in an app-in-progress that I'm working.

Happy to put some more time into this if necessary, but it's a fairly small change that seems like a simple oversight when first written.