Add null option for md-select
You can pass a new attribute to the components md-select called withNullOption
(I don't know if it's the best name for the parameter) and this option would allow
that the prompt can be also a enabled option. This null option it's useful when you
want to clear a <select> and you don't want to force the user to mark one of the
available options.